// autogpt_platform/frontend/src/app/api/__generated__/models/profileDetails.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ProfileDetailsAvatarUrl } from "./profileDetailsAvatarUrl";

/**
 * Marketplace user profile (including read-only fields)
 */
export interface ProfileDetails {
  username: string;
  name: string;
  description: string;
  avatar_url: ProfileDetailsAvatarUrl;
  links: string[];
  is_featured: boolean;
}